Code Working, No errors. Still no output.
The code is working fine, but the author left some intentional errors.
It is up to us to work and fix it(actually nothing is broken, its just doesn't displays the output on the screen)

here is a plan. write a solid strategy in the review and state each requirement clearly. What do we want our code to do and what kind of output do we want form it.