Added exit()
Added exit(0) after the gold_room user input. When the user in the gold room enters the input for the for Line 4 print("this room is full of gold take as much as you want"), as 0 or 1 the message pops up giving user a compliment. There it should stop the game as user has won the game but instead game goes on and keeps asking for the input. So I added the exit(0) after user has won the game.